Now properly breaking lines for commit messages
This is a test of the line breaking feature. If it is working properly,
this line will be correctly broken to be on the same indentation as the
line above. It did not work for the first line unfortunately, but I
think that may have been a result of the comma.

Here is another paragraph testing the functionality of the line breaks.
It appears to be working for this paragraph.